% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQV.R
\name{CountGenesFound}
\alias{CountGenesFound}
\title{Count the number of known genes under a particular QTL}
\usage{
CountGenesFound(qtl_list, trait, trait_gene_list, marker_list)
}
\arguments{
\item{qtl_list}{A matrix containing the chromosome
on which the QTL was originally found ('Chromosome'), the Leftmost
 Confidence interval's location,the Rightmost Conficence interval's location,
 the trait for which the QTL were discovered, and the lengths of the QTL.
 Required columns: (Chromosome, LeftmostMarker, RightmostMarker, Trait, Length)}

\item{trait}{The trait for which you would like to produce a gene list.}

\item{trait_gene_list}{A 4 column matrix containing a list of known genes.
Columns should include "GeneID', "Trait", "Chromosome", and "Locus".}

\item{marker_list}{A 3 column matrix containing the marker loci
(ID, Chromosome, and Base)).

TODO fix doc}
}
\value{
A matrix containing the number of genes identified within a QTL
for a specific trait, as well as the number of QTL that were identified
for that trait
}
\description{
This function determines the number of genes found by a QTL in the original
mapping experiment. Relies on the FilterGeneList function to ensure
the appropriate genes are being used.
}
