% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQV.R
\name{FilterGeneList}
\alias{FilterGeneList}
\title{Create known gene lists appropriate for specific traits and your style of QTL mapping}
\usage{
FilterGeneList(trait, gene_list, marker_list, drop_tandem = TRUE)
}
\arguments{
\item{trait}{The trait for which you would like to produce a gene list.}

\item{gene_list}{A matrix containing a list of known genes. Columns must include "ID', "Trait",
"Chromosome", and "Base".}

\item{marker_list}{A matrix containing the marker loci Required columns: ID, Chromosome, and
Base.}

\item{drop_tandem}{If true, remove "tandem repeats" - multiple genes with no markers between
them.}
}
\value{
A 4 column matrix containing "ID', "Trait", "Chromosome", and "Base" for each remaining
  gene in the trait specified.
}
\description{
This function takes a gene list and refines it to only specific trait you would like to examine.
Will remove duplicate rows. Additionally, the function will act to reduce the final list if
drop_tandem is true, so as to avoid the artificial inflation of identified genes that linkage
causes. This reduction is performed such that a single randomly chosen gene from any set of genes
that lie between two consecutive markers remains.
}
