% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQV_1_3_20.R
\name{Gene_List_Groomer}
\alias{Gene_List_Groomer}
\title{Create known gene lists appropriate for specific traits and
your style of QTL mapping}
\usage{
Gene_List_Groomer(MarkerList, gene_list, Placement_Type, Trait)
}
\arguments{
\item{MarkerList}{A 3 column matrix containing the marker loci
(ID, Chromosome, and Base)).}

\item{gene_list}{A 4 column matrix containing a list of known genes.
Columns should include "GeneID', "Trait", "Chromosome", and "Locus".}

\item{Placement_Type}{Either 'extension' or 'centered'. This depends on the
mapping style of the original QTL experiment.}

\item{Trait}{The trait for which you would like to produce a gene list.}
}
\value{
A 4 column matrix containing "GeneID', "Trait", "Chromosome",
and "Locus" for each remaining gene in the trait specified.
}
\description{
This function takes a gene list and refines it based on the specific trait you
would like to examine. Will remove duplicate rows. Additionally, the function
will act to reduce the final list in the case of QTL mapping that involved extension of the 95% confidence
intervals to the nearest marker, so as to avoid the artificial inflation of
identified genes that linkage causes. This reduction is performed such that a single
random gene from any set of genes that lie between two consecutive markers remains.
}
