% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQV_1_3_20.R
\name{GeneCounter}
\alias{GeneCounter}
\title{Count the number of known genes under a particular QTL}
\usage{
GeneCounter(QTL_with_Metadata, gene_list, Trait, Placement_Type, MarkerList)
}
\arguments{
\item{QTL_with_Metadata}{A 7 column matrix containing the chromosome
on which the QTL was originally found ('Chromosome'), the Leftmost
 Confidence interval's location,the Rightmost Conficence interval's location,
 the trait for which the QTL were discovered, the treatment that was applied,
 the mapping method, the overall type of experiment, and the lengths of the QTL.}

\item{gene_list}{A 4 column matrix containing a list of known genes.
Columns should include "GeneID', "Trait", "Chromosome", and "Start_Site".}

\item{Trait}{The trait for which you would like to produce a gene list.}

\item{Placement_Type}{Either 'extension' or 'centered'. This depends on the
mapping style of the original QTL experiment.}

\item{MarkerList}{A 3 column matrix containing the marker loci
(ID, Chromosome, and Base)).}
}
\value{
A matrix containing the number of genes identified within a QTL
for a specific trait, as well as the number of QTL that were identified
for that trait
}
\description{
This function determines the number of genes found by a QTL in the original
mapping experiment. Relies on the Gene_List_Groomer function to ensure
the appropriate genes are being used.
}
