% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQV.R
\name{GeneFoundLikelihood}
\alias{GeneFoundLikelihood}
\title{TODO write docs}
\usage{
GeneFoundLikelihood(
  gene_chr,
  gene_locus,
  qtl_ext_length,
  placement_type,
  per_marker_likelihood,
  sectioned_marker_list,
  chromosome_size
)
}
\arguments{
\item{gene_chr}{TODO}

\item{gene_locus}{TODO}

\item{qtl_ext_length}{TODO}

\item{placement_type}{TODO}

\item{per_marker_likelihood}{TODO}

\item{sectioned_marker_list}{TODO}

\item{chromosome_size}{TODO}
}
\value{
TODO
}
\description{
TODO write docs
}
