% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQV_1_3_20.R
\name{Marker_Sectioner}
\alias{Marker_Sectioner}
\title{Create chromosome specific matrices of QTL mapping marker locations}
\usage{
Marker_Sectioner(number_chromosomes, MarkerList)
}
\arguments{
\item{number_chromosomes}{Numeric value refering the the number
of chromosomes in the organism of interest}

\item{MarkerList}{A 3 column matrix containing the marker loci
(ID, Chromosome, and Base)).}
}
\value{
N matrices of marker loci

Sectioned_List, a list of the names of the N matrices of marker loci
}
\description{
This function produces N matrices containing the locations of the
markers used in a QTL mapping experiment, where N refers to the number
of chromosomes in the organism of interest.
}
