% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQV_1_3_20.R
\name{QTL_Placement_Probabilities}
\alias{QTL_Placement_Probabilities}
\title{Determine the probability of placing a particular QTL on any marker in the genome}
\usage{
QTL_Placement_Probabilities(
  QTLofInterest,
  chromosome_size,
  Sectioned_List,
  Placement_Type
)
}
\arguments{
\item{QTLofInterest}{A matrix containing (at minimum) the chromosome on which the QTL
was originally found ('Chromosome'), the Leftmost Confidence interval's location,
and the Rightmost Conficence interval's location.}

\item{chromosome_size}{A matrix containing the number of each chromosome,
their total lengths, and the first and last markers used for the original
mapping experiment.}

\item{Sectioned_List}{The list of the names of the N matrices of marker loci
resulting from use of the Marker_Sectioner function.}

\item{Placement_Type}{Either 'extension' or 'centered'. This depends on the
mapping style of the original QTL experiment.}

\item{MarkerList}{A 3 column matrix containing the marker loci
(ID, Chromosome, and Base)).}
}
\value{
A list of the probability of placing each QTL on any individual marker
in the genome
}
\description{
This function determines the likelihood of placing a QTL on every individual
marker in the genome. It takes into account the sizes of your identified QTL,
the lengths of each chromosome, the locations of markers, and the type of mapping
that originally produced the QTL.
}
