% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQV.R
\name{QTLPlacementProbabilities}
\alias{QTLPlacementProbabilities}
\title{Determine the probability of placing a particular QTL on any marker in the genome}
\usage{
QTLPlacementProbabilities(
  qtl_list,
  placement_type,
  sectioned_marker_list,
  chromosome_size
)
}
\arguments{
\item{qtl_list}{A matrix containing (at minimum) the chromosome on which the QTL was originally
found ('Chromosome'), the Leftmost Confidence interval's location, and the Rightmost Conficence
interval's location.}

\item{placement_type}{Either 'extension' or 'centered'. This depends on the mapping style of the
original QTL experiment.}

\item{sectioned_marker_list}{The list of the names of the N matrices of marker loci resulting
from use of the Marker_Sectioner function.}

\item{chromosome_size}{A matrix containing the number of each chromosome and the first and last
markers used for the original mapping experiment.}
}
\value{
A list with length nrow(qtl_list), representing the probability of choosing any
  particular marker reachable with that QTL
}
\description{
This function determines the likelihood of placing a QTL on any individual marker in the genome
that is reachable by that QTL. If placement_type is "extension", the likelihood reflects the
probability of choosing that marker and either L->R or R->L mapping. It takes into account the
sizes of your identified QTL, the lengths of each chromosome, the locations of markers, and the
type of mapping that originally produced the QTL.
}
