% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQV.R
\name{SectionMarkers}
\alias{SectionMarkers}
\title{Create chromosome specific matrices of QTL mapping marker locations}
\usage{
SectionMarkers(marker_list, num_chromosomes)
}
\arguments{
\item{marker_list}{A 3-column matrix containing the marker loci used in a QTL mapping experiment
(ID, Chromosome, and Base)).}

\item{num_chromosomes}{Numeric value refering the the number of chromosomes in the organism of
interest.}
}
\value{
a list of N matrices of marker loci
}
\description{
This function splits marker_list into a list pf matrices for each chromosome.
}
