% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPQV_1_3_20.R
\name{SPQValidate}
\alias{SPQValidate}
\title{Validate QTL using a list of previously known genes}
\usage{
SPQValidate(
  QTL_with_Metadata,
  Trait,
  number_repetitions,
  gene_list,
  chromosome_size,
  MarkerList,
  WholeGenomeGeneDistribution,
  Placement_Type,
  Simulation_Environment
)
}
\arguments{
\item{QTL_with_Metadata}{A 7 column matrix containing the chromosome
on which the QTL was originally found ('Chromosome'), the Leftmost
 Confidence interval's location,the Rightmost Confidence interval's location,
 the trait for which the QTL were discovered, the treatment that was applied,
 the mapping method, the overall type of experiment, and the lengths of the QTL.}

\item{Trait}{The trait for which you would like to analyze your QTL.}

\item{number_repetitions}{An integer defining the number of simulations to run.}

\item{gene_list}{A 4 column matrix containing a list of known genes.
Columns should include "GeneID', "Trait", "Chromosome", and "Locus".}

\item{chromosome_size}{A matrix containing the number of each chromosome,
their total lengths, and the first and last markers used for the original
mapping experiment}

\item{MarkerList}{A 3 column matrix containing the marker loci
(ID, Chromosome, and Base)).}

\item{WholeGenomeGeneDistribution}{A 4 column matrix of all genes in the
genome. Include Chromosome, GeneStart, GeneEnd, and GeneMiddle. The SPQValidate
function uses GeneMiddle to simulate gene placement.}

\item{Placement_Type}{Either 'extension' or 'centered'. This depends on the
mapping style of the original QTL experiment.}

\item{Simulation_Environment}{A new environment in which a dataframe containing the
results of each simulation is stored. This dataframe can be accessed as
Your_Environment$Simulation_Dataframe.}
}
\value{
A matrix containing the lengths of QTL (as an identifier), the true observed
value for identified gene number, the mean value resulting from the simulations,
and the upper and lower 95% confidence intervals for the simulation distributions.
}
\description{
This function determines the number of genes expected to be identified by a QTL
of a particular length, given the distribution of genes and mapping markers in
the genome. Relies on the Gene_List_Groomer, GeneCounter, MarkerSectioner,
and QTL_Placement_Probabilities functions.
}
